requirements.txt
aws-cdk-lib==2.186.0
constructs>=10.0.0,<11.0.0
pytest==6.2.5
bandit[toml]==1.8.3
ruff==0.11.5
